[OGUI-702] Refactor panels structure
