+++ Control/public/workflow/panels/flps/flpSelectionPanel.js
@@ -13,7 +13,7 @@
 */
 
 import {h} from '/js/src/index.js';
-import pageLoading from '../common/pageLoading.js';
+import pageLoading from './../../../common/pageLoading.js';
 
 /**
  * Create a selection area for all FLPs from consul
