+++ Control/public/workflow/panels/variables/advancedPanel.js
